more refactoring, renames and revert test break
