+++ sdk/communication/communication-sms/test/internal/node/smsClientWithToken.internal.spec.ts
@@ -2,27 +2,28 @@
 // Licensed under the MIT license.
 
 /**
- *  These are a duplicate of the public SmsClientWithToken.spec.ts tests, but with additional logic to enable recording/playback.
+ * ###WORKAROUND###
+ *  This duplicates of public SmsClientWithToken.spec.ts tests, but with additional logic to enable recording/playback.
  *  This is a workaround because Http Requests with Randomized UUIDs do not play well with the recorder
  *  These tests will be skipped in Live Mode since the public tests run in live mode only.
  */
 
 import { env, isLiveMode, isPlaybackMode, record, Recorder } from "@azure/test-utils-recorder";
-import { SmsSendRequest, SmsClient } from "../../../src/smsClient";
-import { assert } from "chai";
+import { SmsClient } from "../../../src/smsClient";
 import { isNode } from "@azure/core-http";
 import * as dotenv from "dotenv";
 import * as sinon from "sinon";
 import { parseConnectionString } from "@azure/communication-common";
-import { createCredential, recorderConfiguration } from "../../utils/recordedClient";
+import { createCredential, recorderConfiguration } from "../../public/utils/recordedClient";
 import { Uuid } from "../../../src/utils/uuid";
 import { Context } from "mocha";
+import sendSmsSuites from "../../public/suites/smsClient.send";
 
 if (isNode) {
   dotenv.config();
 }
 
-describe("SmsClientWithToken [Playback/Record]", async () => {
+describe("SmsClient Using Token Based Authentication [Playback/Record]", async () => {
   let recorder: Recorder;
 
   beforeEach(async function(this: Context) {
@@ -31,11 +32,15 @@ describe("SmsClientWithToken [Playback/Record]", async () => {
       this.skip();
     }
     recorder = record(this, recorderConfiguration);
-    recorder.skip("browser");
     if (isPlaybackMode()) {
       sinon.stub(Uuid, "generateUuid").returns("sanitized");
       sinon.stub(Date, "now").returns(0);
     }
+
+    const credential = createCredential();
+    const endpoint = parseConnectionString(env.AZURE_COMMUNICATION_LIVETEST_CONNECTION_STRING)
+      .endpoint;
+    this.smsClient = new SmsClient(endpoint, credential);
   });
 
   afterEach(async function(this: Context) {
@@ -47,24 +52,5 @@ describe("SmsClientWithToken [Playback/Record]", async () => {
     }
   });
 
-  it("can send an SMS when url and token credential are provided", async function() {
-    const credential = createCredential();
-    const endpoint = parseConnectionString(env.AZURE_COMMUNICATION_LIVETEST_CONNECTION_STRING)
-      .endpoint;
-    const fromNumber = env.AZURE_PHONE_NUMBER as string;
-    const toNumber = env.AZURE_PHONE_NUMBER as string;
-
-    const smsClient = new SmsClient(endpoint, credential);
-
-    const sendRequest: SmsSendRequest = {
-      from: fromNumber,
-      to: [toNumber],
-      message: "test message"
-    };
-
-    const responses = await smsClient.send(sendRequest);
-    const response = responses[0];
-    assert.equal(response.httpStatusCode, 202);
-    assert.isTrue(response.successful);
-  }).timeout(5000);
+  describe("when sending SMS", sendSmsSuites);
 });
