+++ sdk/communication/communication-sms/test/public/smsClient.spec.ts
@@ -5,9 +5,9 @@ import { SmsClient } from "../../src/smsClient";
 import { env, record, Recorder } from "@azure/test-utils-recorder";
 import { isNode } from "@azure/core-http";
 import * as dotenv from "dotenv";
-import { recorderConfiguration } from "../utils/recordedClient";
+import { recorderConfiguration } from "./utils/recordedClient";
 import { Context } from "mocha";
-import smsClientSuites from "./suites/smsClientSuites";
+import sendSmsSuites from "./suites/smsClient.send";
 
 if (isNode) {
   dotenv.config();
@@ -16,18 +16,15 @@ if (isNode) {
 describe("SmsClient [Live]", async () => {
   let recorder: Recorder;
 
-  before(async function(this: Context) {
-    this.smsClient = new SmsClient(env.AZURE_COMMUNICATION_LIVETEST_CONNECTION_STRING as string);
-  });
-
   beforeEach(async function(this: Context) {
     recorder = record(this, recorderConfiguration);
-    recorder.skip(undefined); // only run in live mode
+    recorder.skip(undefined, "A UUID is randomly generated within the SDK and used in the HTTP request and cannot be preserved.");
+    this.smsClient = new SmsClient(env.AZURE_COMMUNICATION_LIVETEST_CONNECTION_STRING as string);
   });
 
   afterEach(async function(this: Context) {
     await recorder.stop();
   });
 
-  describe("when sending SMS", smsClientSuites);
+  describe("test send method", sendSmsSuites);
 });
